Stop appending fragmentainer groups when flow thread offset approaches infinity.
The final column height is a function of the difference between the logical
bottom and logical top of the flow thread portion of a given fragmentainer
group. If the logical top is LayoutUnit::max(), we know for sure that the
bottom won't be any larger than that. Just give up in such cases (and keep
using the current fragmentainer group), rather than ending up dividing by zero.

BUG=703543

Review-Url: https://codereview.chromium.org/2784493002
Cr-Original-Commit-Position: refs/heads/master@{#460284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ade7ad316d84cfceaf7980c4d14bd79aaea0ebc