Add canUpdateLayout & canHandleGestureEvent to WebViewClient.
This work is in preperation for work to make it illegal to pass a nullptr
for WebViewClient when instantiting a WebView object.

There are a handful of places in WebViewImpl where the logic changes based
on the nullness of WebViewClient. These changes will preserve this logic when
a non-null WebViewClient is supplied.

This include:
- Returning WebInputEventResult::NotHandled from handleGestureEvent if the
  client specifically opts out of handling the event.
- Not processing layout updates if the client opts out.

After removing null checks in WebViewImpl the requirement for these new methods
are to be reviewed and they should be removed if possible.

Also, in preparation for forcing clients to pass a non null WebViewClient
I made the dtor for WebViewClient public so it can be instantiated without
derivation.

BUG=696895

Review-Url: https://codereview.chromium.org/2762553002
Cr-Original-Commit-Position: refs/heads/master@{#460298}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a6375e628f5405f33f9061ee25644eaa90a81c2