Enable transmitting Fetch Requests over Mojo
This CL implements support for sending Fetch API requests over Mojo. The
used type in Blink is WebServiceWorkerRequest, where the browser side
will use ServiceWorkerFetchRequest.

It's a lot of extra code and definitely more duplication. However, this
is something that we'll need in the near future to support Onion Soup of
the other users of these types, too. There also are opportunities to
remove duplication, notably of the enum values, by adopting the Mojo
enumerations as opposed to the ones in WebURLRequest.

The introduced type now also supersedes the native struct definition
included in service_worker_event_dispatcher.mojom.

BUG=692535

Review-Url: https://codereview.chromium.org/2762303002
Cr-Original-Commit-Position: refs/heads/master@{#460359}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f05540d87c83d707b4e9d7a00f29ebd423b9cfb8