Media: Delete Pause-To-Buffer.
This code was broken in M52, but nobody has complained, so remove it.

Pause-To-Buffer is a consistent source of surprise, and mistakes leading
to accidental bandwidth consumption. The new download button is a much
better interface (at least in cases where it appears).

Review-Url: https://codereview.chromium.org/2689323002
Cr-Original-Commit-Position: refs/heads/master@{#460563}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f5bb6155b40ff1533e4310f18e294ab34b55f14