Simplify the id fast path for non-right most selectors in querySelector.
The id path inside SelectorQuery::findTraverseRootsAndExecute is not actually
reachable for the right most selector since we handle that separately inside
the top level ::execute() call. We also only ever take the fast path when
rootNode.isConnected() (as checked by canUseFastPath) so we can use that too.

By using these two truths inside this code we can simplify what it was doing
and delete a bunch of the code.

BUG=703900

Review-Url: https://codereview.chromium.org/2786763002
Cr-Original-Commit-Position: refs/heads/master@{#460572}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3b15e0c4bb54971c3d689537a2b08c91a89238b