Move registration of LocalFrame mojo interfaces to ModulesInitializer.
To enable this, added a method to register initilization callbacks for
LocalFrame initization. This method adds the callbacks to a static vector, and
the callbacks are executed when a new LocalFrame is initizialized.

This is part of the work to remove the dependency between WebLocalFrameImpl and
modules/*.

BUG=703405

Review-Url: https://codereview.chromium.org/2783543004
Cr-Original-Commit-Position: refs/heads/master@{#460684}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94c0ec2f5e66cf6939523e14cf4db4d3267d1ef1