Get rid of computeVisibleSelectionInDOMTreeDeprecated() from WebViewImpl::selectionBounds()
This patch gets rid of |computeVisibleSelectionInDOMTreeDeprecated()| by moving
|updateStyleAndLayoutIgnorePendingStylesheets()| before using |VisibleSelection|
and checking DOM based empty-selection before updating layout.

When selection is empty in DOM tree, e.g. selection.base().isNull(), it can not
be non-null on |VisibleSelection|.

This patch is a part of workaround of [1] for merging.

[1] http://crbug.com/704763 Opacity animation consuming main thread resources

BUG=698633, 704763
TEST=n/a; no behavior changes

Review-Url: https://codereview.chromium.org/2784243002
Cr-Original-Commit-Position: refs/heads/master@{#460710}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8202d740b1c09787cdd3e11d3d15e7211b88f15e