Add an include-paths only jpeg group
Because libyuv switched to compiling with clang for win64 builds:
https://chromium-review.googlesource.com/c/446667/
we ended up with libjpeg being compiled an additional time, also with
clang. Having two copies is undesirable, so we need a way for libyuv to
request the include paths for the appropriate version of libjpeg. This
sets up a group that libyuv can use for that purpose.

Getting libyuv to use this group is a separate change, in the libyuv
repo.

R=brettw@chromium.org
BUG=706627

Review-Url: https://codereview.chromium.org/2779393002
Cr-Original-Commit-Position: refs/heads/master@{#460889}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae920018eb44986c5c9cdf16285361720dd0df59