Move all of csswg-test to wpt/css and temporarily disable all css tests.
I intend to directly land this CL, so what's shown in Rietveld isn't a reflection of the true contents of the patch, which changes more than 6000 files.

This commit will include:
 An import of just wpt/css from wpt@a810200e6e1b0b98527bd1fd791fa110646cfb91
 Removal of external/csswg-test
 Updates to the expectation files to temporarily disable the tests

NOEXPORT=true
TBR=kojii
TBR_REASON=preparing for follow-up CLs which enable the tests; I expect this CL to be low-risk since it disables all of the affected tests.
BUG=706118

Review-Url: https://codereview.chromium.org/2793583002 .
Cr-Original-Commit-Position: refs/heads/master@{#461177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aef3329cadaa6cc1a8d2f558cd33cce6caa95278