Return '' for overflow if overflow-x and overflow-y are different.
According to the CSSOM spec [1], when overflow-x and overflow-y
are different, the computed value of overflow should be the empty
string since it cannot be expressed in the grammar (step 1.2.).

However, we currently return the max of overflow-x and overflow-y
in terms of their enum values, which does not follow the spec.
Firefox handles this issue correctly.

This patch fixes this issue by return an empty string if the two
overflows are different.

[1] https://drafts.csswg.org/cssom/#serialize-a-css-value
[2] https://github.com/w3c/csswg-drafts/issues/1100

BUG=701235

Review-Url: https://codereview.chromium.org/2752623002
Cr-Original-Commit-Position: refs/heads/master@{#461352}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02c25323de5c3f16bb47d5dc6fc6c7b9a939b784