Split StyleSurroundData::offset into separate members.
In generating ComputedStyle, every generated data member must
correspond to a CSS property (whether long hand or short hand).
StyleSurroundData is part of ComputedStyle and we would like to generate
it as well. Unfortunately, the 'left', 'right', 'top' and 'bottom' CSS
properties are grouped as one data member 'offset', which does not
actually correspond to either a long or short hand CSS property, unlike
'margin' and 'padding'. Thus, we cannot generate this data member
directly.

This patch splits 'offset' into 'left', 'right', 'top' and 'bottom'
so that every data member in StyleSurroundData corresponds to a CSS
property. This allows us to generate it in a later patch.

BUG=628043

Review-Url: https://codereview.chromium.org/2705143002
Cr-Original-Commit-Position: refs/heads/master@{#461356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 542dc4c9f70f7c9e1a3ec096c6b62bceceb5e47f