Replace V8HiddenValue methods in generated code with V8PrivateProperty methods.
This CL replaces V8HiddenValue with V8PrivateProperty, but it keeps to use dynamic
strings as property keys.  (It is possible to list all currently used properties and cache
them in V8PrivateProperty, but it is not practical for future updates.)

As far as I investigate, a property key "state" was the only key that was used in both
generated code and hand written code (V8PopStateEventCustom.cpp).
So this CL also updates the .cpp file to share the property symbol.

BUG=611864

Review-Url: https://codereview.chromium.org/2794453002
Cr-Original-Commit-Position: refs/heads/master@{#461388}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 330bfd236aa4574a1ee370d4b369e755c6fe034d