Ensure baseComputedStyle optimisation is cleared during registered custom property animations
This change to StyleResolver ensures that the cached baseComputedStyle
on an animated element is removed while custom property animations are
active. This prevents stale baseComputedStyles from persisting and
failing the equality assertion in ElementAnimations::updateBaseComputedStyle().

The regression test for this bug needed to be a SimTest as it depended
on style recalcs being triggered by animations and not Javascript.
It's not possible to do this deterministically from Javascript as it
requires control over the frame times.

As part of making the SimTest ElementAnimation::animate() and
PropertyRegistration::registerProperty() required changes to reduce their
V8 dependencies. Additionally a test-only method was added to
AnimationClock to turn off generating synthetic frame times.

BUG=701250

Review-Url: https://codereview.chromium.org/2757543002
Cr-Original-Commit-Position: refs/heads/master@{#461423}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 767e438649154d1985a803c93111e0860a29fa2f