Re-land: Initial skeleton of Accessibility Object Model Phase 1
Landed: r460670   - https://codereview.chromium.org/2750533006/
Reverted: r461332 - https://codereview.chromium.org/2796553002

See the diff between patch set 1 and patch set 2 for the fix to the crash
that the initial patch caused.

Explainer: https://github.com/WICG/aom/blob/master/explainer.md
Spec: https://wicg.github.io/aom/spec/

This change adds an accessibleNode getter on Element, and implements
the role and label properties of AccessibleNode, including a LayoutTest
of the supported functionality.

Properties on an AccessibleNode are stored in a map for efficiency
since there are dozens of possible properties and few will need to be
set on any one particular object.

In existing accessibility code, places where we previously retrieve an
ARIA attribute are replaced with a new wrapper that first checks the
ARIA attribute and then checks the equivalent AOM property.

Once this change lands it should be relatively straightforward to implement
the rest of the properties of AccessibleNode and complete Phase 1.

BUG=680345,707218
TBR=haraken,aboxhall,esprehn

Review-Url: https://codereview.chromium.org/2791943002
Cr-Original-Commit-Position: refs/heads/master@{#461438}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46fca1f2e467228e197bb0c792450ad54667bd5e