Remove calls to SetNeedsCommit and SetNeedsUpdateLayers in RenderWidget
This finds a bunch of redundant or incorrect calls to SetNeedsCommit
and SetNeedsUpdateLayers in RenderWidget and RenderViewImpl. It removes
APIs from RenderWidget and RenderWidgetCompositor when possible. Some
are converted to setNeedsBeginFrame() (aka SetNeedsAnimate) to trigger
a main frame, because it's unclear if that's needed to me. Others are
just removed.
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel
TBR=pdr

Review-Url: https://codereview.chromium.org/2787203003
Cr-Original-Commit-Position: refs/heads/master@{#461752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87fbafb0be4bf8ccc0cd5fa2578527c0aa88b704