Make ComputedStyle generation deterministic.
Since Python dicts are unordered, it is possible that, when we generate
ComputedStyle with no changes to CSSProperties, the generated code is
different (e.g. different order of enum declarations). This is bad
when we start showing diffs for generated code.

This patch enforces an ordering on the generate code so that the
generated code should always be the same given the same input. Namely:

- Enum type declarations are ordered in alphabetical order by name.
- Fields are ordered first by size (for packing), and then in
  alphabetical order by name when there's a tie.

BUG=628043

Review-Url: https://codereview.chromium.org/2686473004
Cr-Original-Commit-Position: refs/heads/master@{#461956}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe7b8a3b74de1097ffc1ff819c6b2109dcb6800b