CompositorWorkerProxyClientImpl to CompositorProxy cross thread reference is using WeakMember
Because of a bug in Oilpan causing DCHECK to be skipped, we failed to detect that CompositorWorkerProxyClientImpl to CompositorProxy reference isn't using CrossThreadWeakPersistent.

The DCHECK bug will be fixed in https://codereview.chromium.org/2773013002/ .

This CL splits out the CompositorProxy registration to a new class so we don't need to put CrossThreadWeakPersistents in a HashMap.

BUG=705872

Review-Url: https://codereview.chromium.org/2774223002
Cr-Original-Commit-Position: refs/heads/master@{#461972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0af77d69187ee8a529072cb3995bd111aaed5864