Image Capture: remove MediaSettingsRange.current
This CL removes MediaSettingsRange.current member, which
was never in the Spec to start with. "Current" feature values
still need to be kept around for getMediaTrackSettings(): with
the |current| member gone, we aneed to dd a new member
|m_settings| to keep them (and it encompasses the previous
points of interest vector).

LayoutTests updated accordingly.

BUG=700607, 707262

Review-Url: https://codereview.chromium.org/2795923003
Cr-Original-Commit-Position: refs/heads/master@{#461974}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1052a48bc85cc5d68c0f7dc0782b577e92cc9c7