Add support for rebaselining for a port from a non-corresponding builder.
This CL changes TestBaselineSet to store port names as well as builds,
and makes it so that the rebaseline method, when given constructing a
list of commands to invoke for a given TestBaselineSet, the port name
is according to the TestBaselineSet and doesn't necessarily need to
correspond to the builder name.

After this change, I believe that the only change left before
rebaseline-cl supports rebaselining with only some try bot results
available is to change the way that TestBaselineSets are created
in rebaseline_cl.py.

BUG=673966

Review-Url: https://codereview.chromium.org/2790863002
Cr-Original-Commit-Position: refs/heads/master@{#461980}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 339734d88bb8df1cc19b7f0bee7cdcf0645fb1ef