Introduce the nested namespace ::blink::cssvalue, start with CSSColorValue.
This patch is the first in a series that will move all of the classes in the range CSS.*Value from the ::blink namespace to the ::blink::cssvalue namespace. It will be followed by work to introduce the ::blink::cssom namespace.

The original design doc for this work is linked in the attached bug. Note that this patch does not rename the class or move the file to a new subdirectory. The immediate goal is to prevent name collisions cssvalue and cssom types - there is no need to re-organize or rename the files to achieve this basic goal.

BUG=667961

Review-Url: https://codereview.chromium.org/2796183002
Cr-Original-Commit-Position: refs/heads/master@{#462010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80cc48147b27aa364144e4b6d49524fb111274a9