Workaround for Samsung email issues
Samsung's email works around Range.getClientRects()'s behavior to return
an empty sequence for a collapsed range, and the workaround is to insert
a non-zero space, and call Range.getClientRects(), and then remove the
non-zero space. With the changes introduced in
https://codereview.chromium.org/2687273002, now inserting a non-zero
space triggers a selectionchanged event and it causes an infinite loop,
leading to issues with scroll, caret blink, and paste.

Because it is difficult for Samsung to update their email any time soon,
we undo the above changes only for WebView by adding a runtime
flag and set it. The rationale for generalizing this to WebView is:

1) Package name checks generally don't work very well - there may be
variants for different devices or beta channels.
2) JS library developers should still be concerned about out-of-date
WebView libraries anyways.

This should be fixed for version N and above, so we
limit the workaround to Android M and below.

BUG=698752, 699943

Review-Url: https://codereview.chromium.org/2776073002
Cr-Original-Commit-Position: refs/heads/master@{#462173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e54ccdb9fbd84362cba3c5858deaad215e572f5c