Make WebScrollbarThemePainter paint at the origin
Currently many paint rects for WebScrollbarThemePainter are
not located at the origin.

This is being changed because it's currently blocking landing
https://codereview.chromium.org/2768143002, which reimplements
SkPictureRecorder / SkPicture with new classes in Chromium.
SkPictureRecorder has private internal behavior to support
non-origin canvases that the new recording code does not.
Additionally, the skia team would like to remove dependencies
on non-origin recording.

Alternatives to this patch include modifying WebScrollbarThemeGeometry
to not attempt to paint scrollbars at their offset on the page
(e.g. track rects at (785, 0, 15, 600) for an 800x600 page), but this
is a lot more involved.  Another alternative is to modify
GraphicsContextCanvas to not need to know about device clips and
to deal with rasterizing into cg in local space.

This seems like the smallest and safest change to allow the above
patch to land and is in line with skia team desires.

Change-Id: Iad7b9fb7051cfe6c9242677ef807aaf1ae990661
Reviewed-on: https://chromium-review.googlesource.com/469073
Reviewed-by: Chris harrelson <chrishtr@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#462552}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 029eaee6dca104df967d42cf6bfbb40d37443e25