Add DocumentMarker::MatchStatus enum for text match markers
Suggested by @yosin here:
https://codereview.chromium.org/2801483002#msg20

Modify the DocumentMarker() constructor for text match markers and DocumentMarkerController::addTextMatchMarker() to take an enum for whether or not the text match is active instead of a bool. This will make it more clear in call sites what the parameter means.

BUG=707867

Review-Url: https://codereview.chromium.org/2801483002
Cr-Original-Commit-Position: refs/heads/master@{#462303}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c198feadcc54ea92134df2dc3a5f4f390fcd9493