Generalise 'flag' template to 'primitive' when generating ComputedStyle.
Currently we have 'flag' fields which represent boolean fields inside
ComputedStyle. We can generalise this to not only booleans, but any
primitive value such as int or float. This patch renames the 'flag'
template to 'primitive' and adds a special case for 'bool' fields which
allows them to be packed as a single bit.

This is prework for generating fields that are stored as ints or floats.

This patch does not affect generated code.

BUG=628043

Review-Url: https://codereview.chromium.org/2793843002
Cr-Original-Commit-Position: refs/heads/master@{#462344}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e8890d72e9824fce593382d5eb23210e4d7db177