Avoid duplicate functions/code in core/editing: DirectionalSelection
While experimenting with unity builds I encountered a few duplicate
symbols and functions in core/editing. This patch renames, moves
and unifies them.

shouldAlwaysUseDirectionalSelection is a common helper function
and since it's not identicallty implemented everywhere it cannot
be merged. Instead make the one in SelectionModifier a member function
to move it out of global scope.

BUG=708949
R=yosin@chromium.org

Review-Url: https://codereview.chromium.org/2798143003
Cr-Original-Commit-Position: refs/heads/master@{#462427}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0736e37b73d70fc56a93eeac32fff4a56c946c2d