Avoid duplicate functions/code in core/editing: MatchResultICU
While experimenting with unity builds I encountered a few duplicate
symbols and functions in core/editing. This patch renames, moves
and unifies them.

There is a global MatchResult in css/SelectorChecker.h and if that one
is included anywhere in editing, then it will collide with a local
MatchResult used in iterators. Renaming the local one MatchResultICU
will both match the name of other symbols and avoid the collision.

BUG=708949
R=yosin@chromium.org

Review-Url: https://codereview.chromium.org/2806433002
Cr-Original-Commit-Position: refs/heads/master@{#462431}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2bd991090f462a64efe0f6b26bf0ad35e19386b0