Update Error Event inside a worker to provide the exact exception value
In WorkerOrWorkletScriptController.cpp, 'ErrorEvent' used to be created
without exception value, so the exception, |m_error|, is always set to
null. A new constructor of ErrorEvent is added in this patch which
takes exception as input and set |m_error| to the exception value.

TEST=error-script.js and error-worker-script.js

BUG=640724

Review-Url: https://codereview.chromium.org/2804533002
Cr-Original-Commit-Position: refs/heads/master@{#462458}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0662e1de8c6d9800c4baed479249c1fa5201c24c