Resume() should return if context is already running
It's valid to call AudioContext.resume() if the context is already
running.  The promise should resolve immediately; we can handle the
promise now instead of saving it on a list to be resolved later.  This
prevents us from setting the context state to running again when it's
already running, because that triggers a DCHECK.

The spedc also says to resolve the promise when the context is already
running.

BUG=707949
TEST=audiocontext-suspend-resume.html

Review-Url: https://codereview.chromium.org/2798973002
Cr-Original-Commit-Position: refs/heads/master@{#462506}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5149859e198b741fe8c89a165e4035339efcdb62