Add -f flag to git cl upload when uploading patchset in wpt importer
The last couple import jobs failed when uploading the second patchset;
I think that it failed at this point (and not when uploading the
CL initially) because when initially uploading the CL it uses -f.

We want to use -f in general to ignore presubmit warnings for things
like tabs.

Review-Url: https://codereview.chromium.org/2802093002
Cr-Original-Commit-Position: refs/heads/master@{#462698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65b779860c2fbcda49066ea532c5cb3178fdc65e